Split utils file
