+++ src/base/base-component.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils';
+import {createElement} from '../utils/render';
 
 /**
  * Representing base component interface.
