+++ src/utils/common.js
@@ -1,29 +1,4 @@
-import BaseComponent from './base/base-component';
-import {RenderPosition, MillisecondsEnum} from './mock/consts';
-
-/**
- * Render HTML string to container.
- *
- * @param {HTMLElement|BaseComponent} container
- * @param {BaseComponent} component
- * @param {string} [position=beforeend]
- */
-const render = (container, component, position = RenderPosition.BEFORE_END) => {
-  if (container instanceof BaseComponent) {
-    container.getElement().insertAdjacentElement(position, component.getElement());
-  } else {
-    container.insertAdjacentElement(position, component.getElement());
-  }
-};
-
-/**
- * Create HTMLElement from template string.
- *
- * @param {string} template
- *
- * @return {HTMLElement}
- */
-const createElement = (template) => document.createRange().createContextualFragment(template).firstChild;
+import {MillisecondsEnum} from '../mock/consts';
 
 /**
  * @param {Array} array
@@ -154,11 +129,9 @@ const getEventDurationString = (diff) => {
 };
 
 export {
-  render,
   getEventPlaceholder,
   shuffleArray,
   getRandomNumber,
-  createElement,
   getEventDurationString,
   getRandomDescription,
   groupEventsByDays,
